pam_mount: allow luserconf outside home directory
[It would seem luserconf was once intended to be used for mounting SMB/NFS
shares _in addition to an always-present_ home directory. –jengelh]

When mounting the home directory [itself] via pam_mount, it is
[necessary] to keep the user's pam_mount.conf.xml elsewhere. It is
particularly helpful if the home directory is encrypted.

This patch recognizes an absolute path in <luserconf> and, if so,
does not prepend the home directory.

It also replaces the various placeholders, such as %(USER) and "~" in
<luserconfig>. It allows all users to keep personal config files in
similar locations.